+++ python/GafferSceneUI/InteractiveRenderUI.py
@@ -44,6 +44,8 @@ import GafferScene
 import GafferUI
 import GafferImageUI
 
+import IECore
+
 ##########################################################################
 # ImageView UI for the viewed image's render node (if available)
 ##########################################################################
@@ -88,6 +90,7 @@ class _ViewRenderControlUI( GafferUI.Widget ) :
 				GafferUI.Spacer( imath.V2i( 1, 1 ), imath.V2i( 1, 1 ) )
 				self.__label = GafferUI.Label( "Render" )
 				self.__stateWidget = _StatePlugValueWidget( None )
+				self.__messagesWidget = _MessageSummaryPlugValueWidget( None )
 
 		self.__view = view
 
@@ -105,6 +108,7 @@ class _ViewRenderControlUI( GafferUI.Widget ) :
 
 			if not statePlug.isSame( self.__stateWidget.getPlug() ) :
 				self.__stateWidget.setPlug( statePlug )
+				self.__messagesWidget.setPlug( renderNode["messages"] )
 				self.__renderNodePlugDirtiedConnection = renderNode.plugDirtiedSignal().connect( Gaffer.WeakMethod( self.__renderNodePlugDirtied ) )
 
 			# We disable the controls if a render is in progress, but not being viewed
@@ -117,9 +121,11 @@ class _ViewRenderControlUI( GafferUI.Widget ) :
 				"" if controlsEnabled else "Controls disabled because image is not the one currently rendering."
 			)
 			self.__stateWidget.setEnabled( controlsEnabled )
+			self.__messagesWidget.setEnabled( controlsEnabled )
 
 		else :
 			self.__stateWidget.setPlug( None )
+			self.__messagesWidget.setPlug( None )
 			self.__renderNodePlugDirtiedConnection = None
 
 	@staticmethod
@@ -229,6 +235,105 @@ class _StatePlugValueWidget( GafferUI.PlugValueWidget ) :
 	def __stopClicked( self, button ) :
 		self.getPlug().setValue( GafferScene.InteractiveRender.State.Stopped )
 
+###############################################################################
+# A widget presenting a summary of the messages in a render nodes messages plug
+###############################################################################
+
+class _MessageSummaryPlugValueWidget( GafferUI.PlugValueWidget ) :
+
+	def __init__( self, plug ) :
+
+		self.__summaryWidget = GafferUI.MessageSummaryWidget(
+			displayLevel = IECore.MessageHandler.Level.Warning,
+			hideUnusedLevels = True,
+			buttonToolTip = "Click to open the render log"
+		)
+
+		GafferUI.PlugValueWidget.__init__( self, self.__summaryWidget, plug )
+
+		self.__summaryWidget.levelButtonClickedSignal().connect( Gaffer.WeakMethod( self.__levelButtonClicked ), scoped = False )
+
+		self._updateFromPlug()
+
+	def getToolTip( self ) :
+
+		# Suppress the default messages tool-tip.
+		return ""
+
+	@GafferUI.LazyMethod( deferUntilPlaybackStops = True )
+	def _updateFromPlug( self ) :
+
+		if self.getPlug() is not None :
+			with self.getContext() :
+				messages = self.getPlug().getValue().value
+		else :
+			messages = Gaffer.Private.IECorePreview.Messages()
+
+		self.__summaryWidget.setMessages( messages )
+		self.__summaryWidget.setEnabled( self.getPlug() is not None )
+
+	def __levelButtonClicked( self, level ) :
+
+		window = _MessagesWindow.acquire( self.getPlug() )
+		window.setVisible( True )
+		window.messageWidget().scrollToNextMessage( level )
+
+###############################################################################
+# A utility window containing a render nodes message log
+###############################################################################
+
+## TODO: This is awefully similar to numerous color picker windows, etc...
+# we ideally could do with a GafferUI.PlugWindow or similar.
+class _MessagesWindow( GafferUI.Window ) :
+
+	# Use acquire instead to retrieve an existing window or create a new one
+	def __init__( self, parentWindow, plug ) :
+
+		GafferUI.Window.__init__( self, borderWidth = 8 )
+
+		self.setChild( _MessagesPlugValueWidget( plug ) )
+		self._qtWidget().resize( 600, 500 )
+
+		parentWindow.addChildWindow( self, removeOnClose = True )
+
+		node = plug.node()
+		scriptNode = node.scriptNode()
+		while node and not node.isSame( scriptNode ) :
+			node.nameChangedSignal().connect( Gaffer.WeakMethod( self.__updateTitle ), scoped = False )
+			node.parentChangedSignal().connect( Gaffer.WeakMethod( self.__destroy ), scoped = False )
+			node = node.parent()
+
+		self.__updateTitle()
+
+	def messageWidget( self ) :
+
+		return self.getChild().messageWidget()
+
+	def __updateTitle( self, *unused ) :
+
+		plug = self.getChild().getPlug()
+		self.setTitle( "{} Messages".format( plug.node().relativeName( plug.ancestor( Gaffer.ScriptNode ) ) ) )
+
+	def __destroy( self, *unused ) :
+
+		self.parent().removeChild( self )
+
+	@classmethod
+	def acquire( cls, plug ) :
+
+		script = plug.node().scriptNode()
+		scriptWindow = GafferUI.ScriptWindow.acquire( script )
+		childWindows = scriptWindow.childWindows()
+
+		for window in childWindows :
+			if isinstance( window, cls ) and window.getChild().getPlug().isSame( plug ) :
+				return window
+
+		window = cls( scriptWindow, plug )
+		window.setVisible( True )
+
+		return window
+
 ##########################################################################
 # UI for the messages plug that presents the render log
 ##########################################################################
@@ -254,13 +359,20 @@ class _MessagesPlugValueWidget( GafferUI.PlugValueWidget ) :
 	def getToolTip( self ) :
 
 		# Suppress the default messages tool-tip that otherwise appears all over the log window.
-		return None
+		return ""
+
+	def messageWidget( self ) :
+
+		return self.__messages
 
 	@GafferUI.LazyMethod( deferUntilPlaybackStops = True )
 	def _updateFromPlug( self, *unused ) :
 
-		with self.getContext() :
-			messages = self.getPlug().getValue().value
+		if self.getPlug() is not None :
+			with self.getContext() :
+				messages = self.getPlug().getValue().value
+		else :
+			messages = Gaffer.Private.IECorePreview.Messages()
 
 		self.__messages.setMessages( messages )